cli: drop python-netifaces
netifaces doesn't return the permanent MAC address of physical
interfaces. This is causing issues ever since we start to match
interfaces only by permanent MAC addresses.

We only use it to retrieve the list of interfaces and their MACs so
let's drop it and implement this functionalities in the CLI.

get_interfaces() will simply call "ip link" and return a list with the
interface names.

get_interface_macaddress() will try to get the permanent MAC address and
fallback to the transient address if it can't be found.